fitness must be double

+++ src/main/java/ch/kaiki/nn/genetic/GeneticAlgorithmObject.java
@@ -65,7 +65,7 @@ public abstract class GeneticAlgorithmObject implements IGeneticAlgorithmObject 
      * @return the fitness value for this instance.
      */
     @Override
-    public abstract long getFitness();
+    public abstract double getFitness();
 
     /**
      * This method indicates if this instance is in an immature state yet. In mature statues only, the kaiki.ch.kaiki.nn.genetic
@@ -105,8 +105,8 @@ public abstract class GeneticAlgorithmObject implements IGeneticAlgorithmObject 
      */
     @Override
     public int compareTo(@NotNull IGeneticAlgorithmObject o) {
-        long fitness = this.getFitness();
-        long otherFitness = o.getFitness();
+        double fitness = this.getFitness();
+        double otherFitness = o.getFitness();
         if (fitness > otherFitness) {
             return 1;
         } else if (fitness < otherFitness) {
